Add OpenRepoSQL menu item to launch RepoSQL app
Introduced a new `ToolStripMenuItem` named `OpenRepoSQLToolStripMenuItem` in `frmMain.Designer.vb` to enable launching the RepoSQL application from the menu. This item is now part of the `fileToolStripMenuItem` dropdown.

Implemented an event handler `OpenRepoSQLToolStripMenuItem_Click` in `frmMain.vb` to handle the click event, checking for the existence of `RepoSQL.exe` and providing user feedback through status updates and message boxes. The menu item is declared as a `Friend WithEvents` variable for accessibility within the form's class.